refactor: set correct name
